Fix inaccuracy in comparison of elapsed time (dt)
The inaccuracy is small, but it becomes visible when operating
within tight margins of the `sample_time`, like for example, when
working with a simulated clock in a test environment.

Specifically, `dt` cannot be exactly zero because it is used as
a divider. When it would be zero, it is set to a very small,
conventional value instead, that I've renamed MIN_DT for clarity.

When `dt == MIN_DT` it is really representing the case when it
would be zero. Not taking that value into consideration when
comparing the elapsed time since last computation does effectively
increase `sample_time` by `MIN_DT`. It is not a lot, but it is not
accurate either.

By correcting the comparison, very precise use of `sample_time`
becomes possible. One concrete use case is using simulated time
for testing.