Refactor reduce unnecessary wait time in tests
Unless I'm misunderstanding something, the time necessary
for the system to converge should be stable, for any given
configuration of the PID and a given runtime environment.

I notice that the values 120s and 12s are close enough that
120 might have been a typo? Either way, my testing suggests
that 10s is sufficient for the system to converge on my machine.

However, the reliance on wall time for these tests means that
how fast the code is being executed does affect that value.
Before we fix that, I think that the 20% margin provided by
the existing value of 12s is reasonnable.

On my machine, this change reduces the overall time to run
the tests suite by a factor of five (from ~2m14 to ~26s).