Remove redundant testing option
The `dt` option was added when `time_fn` was not available to
initialize PID instances.

Now that `time_fn` allows the time function to be injected as a
dependency, the `dt` option is redundant. It could be kept around
for backward compatibility, but I'd personally remove it because
IMHO the simpler the code the better.